+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.1.9
+
+* `ScrollablePositionedList`:
+
+  * Store scroll state in page storage to fix
+    [issue #43](https://github.com/google/flutter.widgets/issues/43).
+
+* `LinkedScrollControllerGroup`:
+
+  * Add `LinkedScrollControllerGroup.onOffsetChanged` method that calls the
+    provided callback when the scroll offset of the group changes.
+
 # 0.1.8
 
 * `ScrollablePositionedList`: